// src/types/utm.ts
export interface utmObj {
    utm_source: string;
    utm_medium: string;
    utm_campaign: string;
    utm_content: string;
    utm_id: string;
    utm_term: string;
    fullname: string;
    email: string;
    mobile: string;
    project_name: string;
    projects_code: string;
    cluster_code: string;
    lead_source_name: string;
    web: string;
}